v2.20
Client loading updated.
Scripts updated.
Randoms updated.
Methods updated.
GUI updated.

git-svn-id: http://rsbot-client.googlecode.com/svn/trunk@66 128dfb44-bd8a-652a-bd9a-56b36ba659b7

+++ src/org/rsbot/client/Render.java
@@ -5,13 +5,13 @@ package org.rsbot.client;
  */
 public interface Render {
 
-	float getAbsoluteX1();
+	int getAbsoluteX1();
 
-	float getAbsoluteX2();
+	int getAbsoluteX2();
 
-	float getAbsoluteY1();
+	int getAbsoluteY1();
 
-	float getAbsoluteY2();
+	int getAbsoluteY2();
 
 	int getXMultiplier();
